break wal_service into multiple files
+ misc cleanups
