+++ walkeeper/src/lib.rs
@@ -4,7 +4,9 @@ use std::path::PathBuf;
 use std::time::Duration;
 
 pub mod pq_protocol;
+pub mod replication;
 pub mod s3_offload;
+pub mod send_wal;
 pub mod wal_service;
 
 use crate::pq_protocol::SystemId;
